Add WIP cv layer
