+++ src/employees/cv/utils/types.ts
@@ -1,0 +1,37 @@
+
+export type Qualification = {
+    name: string,
+}
+
+export type Qualifications = {
+    name: string,
+    tags: Qualification[],
+}
+
+export type Project = {
+    starred: boolean,
+    customerName: string,
+    description: string,
+    month_from: string,
+    month_to: string,
+    year_from: string,
+    year_to: string,
+    roles: string[],
+}
+
+export type Publication = {
+    name: string,
+    description: string,
+    url: string,
+}
+
+export type EmployeeCv = {
+    name: string,
+    title: string,
+    email: string,
+    imageUrl: string,
+    summary: string,
+    qualifications: Qualifications[],
+    projects: Project[],
+    publications: Publication[],
+}
